Address stylistic comments
